requirements.txt
opencv-python
numpy
pyzbar
pyttsx3
gitpython